collie: fix a vdi operation race
When we send a vdi operaton to the collie, the following
procedures are executed:

(1) the operation message 'A' is sent to the corosync
(2) __sd_deliver is called with w->msg == 'A'
(3) the completion message 'B' is sent to the corosync
(4) __sd_deliver_done is called with w->msg == 'A'
(5) __sd_deliver is called with w->msg == 'B'
(6) __sd_deliver_done is called with w->msg == 'B'

These must be called in this order.

This patch ensures that (4) is called before (5).

Signed-off-by: MORITA Kazutaka <morita.kazutaka@lab.ntt.co.jp>