change program names
Currently, program names in Sheepdog project are:

sheepdog daemon     : collie
sheepdog admin tool : shepherd

Collie is sheepdog, and shepherd is a job name to take care of sheep.
Sheepdog daemon is running on each node, so in this case, there are
many sheepdogs and one shepherd in the cluster.  But, if we take into
account our project name, there should be many sheeps and one
sheepdog.

This patch changes the names to as follows:

sheepdog daemon     : sheep
sheepdog admin tool : collie

Signed-off-by: MORITA Kazutaka <morita.kazutaka@lab.ntt.co.jp>