sheep: increase the number of threads
The current implementation limits the number of VMs on the same host
machines to the number of worker threads; we can run only 4 VMs for
each host machines.

This patch increases the number of threads (and the number of VMs)
from 4 to 64.  I think this fix is temporary and we need to remove the
limitation about the number of VMs in future.

Signed-off-by: MORITA Kazutaka <morita.kazutaka@lab.ntt.co.jp>