sheep: avoid processing I/O until the target object is recovered
I/O operations fail always if the target objects are not recovered
from the previous node membership change, so such operations should be
delayed until the object is recovered.

This patch sorts the order of recovering objects and makes it easy to
check whether the object is recovered or not.

Signed-off-by: MORITA Kazutaka <morita.kazutaka@lab.ntt.co.jp>