sheep: remove verify_object()
verify_object() was used for /[stored dir]/obj/[epoch]/list, which
contains the list of object IDs.  This patch replaces the function to
the following simpler procedures:

 1. create "list.tmp"
 2. write data to list.tmp
 3. rename "list.tmp" to "list"

rename(2) is an atomic operation, so we can create the file in the all
or nothing way; we don't need verification of it.

Signed-off-by: MORITA Kazutaka <morita.kazutaka@lab.ntt.co.jp>