introduce virtual nodes
Currently, Sheepdog data balancing has some problems:

 - When there are only few physical nodes in the cluster, the
   consistent hash ring becomes sparse and data cannot be balanced
   well.
 - Even if some nodes have a larger disk space, we cannot allocate
   more data to the nodes.

This adds preliminary support for virtual nodes; Sheepdog assigns
multiple virtual nodes to each physical node, and creates a consistent
hash ring with virtual nodes.  The number of virtual nodes are fixed
to 64 in this patch, but we can extend it in future.

This patch changes the map between objects and nodes.  So, we need to
reformat Sheepdog cluster to try this feature, but this is a necessary
change, I think.

Signed-off-by: MORITA Kazutaka <morita.kazutaka@lab.ntt.co.jp>