collie: use exit codes to distinguish between errors
When integrating collie into cluster management scripts, it is useful to
be able to tell the difference between different types of error without
needing to parse human-readable error text. In addition to the standard
EXIT_SUCCESS (0) and EXIT_FAILURE (1) exit conditions, we introduce

  EXIT_SYSFAIL (2) - something is wrong with the cluster or local host
  EXIT_EXISTS (3)  - the object already exists so cannot be created
  EXIT_FULL (4)    - no more space is left in the cluster
  EXIT_MISSING (5) - the specified object does not exist
  EXIT_USAGE (64)  - invalid command, arguments or options

and attempt to return these consistently for all collie commands.

Signed-off-by: Chris Webb <chris@arachsys.com>
Signed-off-by: MORITA Kazutaka <morita.kazutaka@lab.ntt.co.jp>