collie: extend vdi list to allow listing a single VDI by name
On a test cluster with three nodes and five hundred VDIs, the early exit
optimisation introduced here reduces the time to find the size, used, and
shared values from a specific VDI from around 3.4s (for naive grep of collie
vdi list output) to around 2s, similar to a 'collie vdi object' command.

However, is it possible to make this and other VDI operations properly
constant- or log-time in the number of VDIs instead of linear?

Signed-off-by: Chris Webb <chris@arachsys.com>
Signed-off-by: MORITA Kazutaka <morita.kazutaka@lab.ntt.co.jp>