sheep: remove object list file
Sheepdog creates a object list file (*.list) when the epoch is
changed.  This file is used for the response of SD_OP_GET_OBJ_LIST.
But when many nodes are added to the cluster at the same time, it
takes a long time to create a many list files, and SD_OP_GET_OBJ_LIST
can result in a timeout error.

This patch completely removes the object list file from Sheepdog.  In
the response of SD_OP_GET_OBJ_LIST, sheep simply calls readdir() on
the store directory, and lists all the object id stored on the local
node.

Signed-off-by: MORITA Kazutaka <morita.kazutaka@lab.ntt.co.jp>