sheep: put init_store() in the proper place
init_store() calls init_epoch_path(), which relys on the log utility
to do the logging. Howevery, log_init() is called after init_store().
So if we enable debug mode, it will leak some of the debug information
on standard output that otherwise is expected in the log file.

This patch makes logger behave itself.

Signed-off-by: Liu Yuan <tailai.ly@taobao.com>
Signed-off-by: MORITA Kazutaka <morita.kazutaka@lab.ntt.co.jp>