sheep: get vdi bitmap logic right in join phase
For newly added nodes, they cannot get old vdi information that are created
before it joins. This is because in the join phase, we don't get vdi bitmap
right and just get vdi bitmaps from nodes that are in new node's sd_list. So
the correct way is: (as MORITA Kazutaka suggested)

1 call get_vdi_bitmap_from_sd_list() to get bitmaps from ndoe's sd_list.
2 call get_vdi_bitmap_from(msg->from) to get bitmap from the delivery node.
3 call get_vdi_bitmap_from(msg->nodes) to get bitmaps from the old configuration.

Signed-off-by: Liu Yuan <tailai.ly@taobao.com>
Signed-off-by: MORITA Kazutaka <morita.kazutaka@lab.ntt.co.jp>