sheep: teach sheepdog to better recovery the cluster
[Problem]

Currently, sheepdog cannot recovery cluster into full functional state if any node in
the cluster fails to join cluster after shutdown, because of being considered unhealthy
(F.g. the targeted epoch content is corrupted, epoch version mismatch). That is, the cluster
can only get worked again only *if* all the nodes join the cluster successfully.

For 14 nodes in the cluster,

      ==========*===    <--- cluster refuses to work
                ^
                |
          unhealthy node

This is quite awkward. The cluster with many nodes after being shutdowned, we easily meet
the condition that some of nodes are unhealthy that are rejected by the master during join
stage.This patch gives sheepdog some kind of intelligence to deal with unhealthy nodes and
process to recovery when all the nodes alive reach the agreement.

[Design]

This patch add a new concept into sheepdog, the *leave node*. The _leave node_ is the one
that the master checks and finds it unhealthy (unmatched epoch content), so marks it as
'leave node', meaning that it is supposed to leave the cluster. The leave nodes are queued
in the leave list, *only* exist during SD_STATUS_WAIT_FOR_JOIN. All the leave nodes stop
sheep itself automatically after being started.

The key idea for *when* the node reach the agreement is very simple, can be summed up into
one equation:

	nr_nodes_in_epoch == nr_nodes_in_sd_list + nr_nodes_in_leave_list

When this is reached, all the nodes alive in the cluster will begin to recovery and finally
with the epoch incremented by 1, if some other nodes are considered to leave. This is because,
after the cluster recovery, we can then try *re-join* the unhealth nodes and probably succeed!

[Cases]

There is no order imposed to start up the nodes in the cluster. That is, you can start up the
nodes after shutdown in arbitrary order, whether it is good node or bad node (Yes, you cannot
know node status before you try to start it up).

For e.g

         ====** <-- the cluster will recovery with 4 nodes alive with epoch incremented by 1
         ==***= <-- the cluster will recovery with 3 nodes alive with epoch incremented by 1
         =***** <-- the cluster will recovery with 1 nodes alive with epoch incremented by 1

                 |* try to re-join
                 v
         =*==== --> ======  <- we might get 6 nodes alive.

The corner case is that we start one of the bad nodes first before any healthy nodes. If this
happens, all the other nodes are considered 'unhealthy', then the very one node will recovery
after we try to join all the nodes.

         *===== --> the cluster will recovery with 1 nodes alive with epoch incremented by 1
                 ^
                 | from now on, we can re-join the nodes, and will end up with 6 nodes alive.

[Note]

This patch changes join_message layout! So you have to recompile all the sheepdogs in your cluster
before it works out.

We need to modify the join_message layout, because of the following scenario:

         =**=== <-- the last 3 nodes have to need know previous 2 leave nodes information.

When the nodes following leave nodes join, the master is supposed to reply with leave nodes information
via send_join_response().

Signed-off-by: Liu Yuan <tailai.ly@taobao.com>
Signed-off-by: MORITA Kazutaka <morita.kazutaka@lab.ntt.co.jp>