sheep: tame sheep to recover the crash cluster
[Rationale]

Currently, we have to start up the frist failed node or last failed one to
recover the crash cluster (nodes with different epoch histories). This patch
simply remove this disgusting constraint.

This patch addes a new concept in the crash cluster recovery phase: mastership transfer.

The master node in the sheepdog is supposed to reply the nodes' join requests, hence managing
other nodes join. It is the first node in the join stage and the last node in the crashed
cluster.

When we recover the crash cluster, the mastership is transfered one another until the last
failed node is started up. When the node is not the last failed one, one of the two nodes simplely
exits, only one node will be left in the recover stage with the mastership.

After the last failed node is started up, we can join others safely with consistent epoch histories.

With this patch, there is no start-up order imposed for the crash cluster to recover.

As side effect, epoch transfer of leave node is removed. Leave node concept is changed a bit during
crashed cluster recovery, that it is defined as "nodes that are supposed to leave and contained in the
latest epoch".

[Test Cases]

The methods that I have tried to test this idea:

$ for i in 0 1 2; do ./sheep/sheep /store/$i -z $i -p 700$i; sleep 1; done
$ collie/collie cluster format
$ for i in 0 1 2; do pkill -f "sheep /store/$i"; sleep 1; done
$ for i in 1 0 2; do ./sheep/sheep /store/$i -z $i -p 700$i; sleep 1; done
$ for i in 0 2; do ./sheep/sheep /store/$i -z $i -p 700$i; sleep 1; done
$ for i in 0 1 2; do ./collie/collie cluster info -p 700$i; done

...
2011-09-25 11:38:17      5 [192.168.0.4:7000, 192.168.0.4:7001, 192.168.0.4:7002]
2011-09-25 11:38:17      4 [192.168.0.4:7000, 192.168.0.4:7002]
2011-09-25 11:38:17      3 [192.168.0.4:7002]
2011-09-25 11:38:17      2 [192.168.0.4:7001, 192.168.0.4:7002]
2011-09-25 11:38:17      1 [192.168.0.4:7000, 192.168.0.4:7001, 192.168.0.4:7002]
...

And Kazutaka's gorgous test case:

for i in 0 1; do ./sheep/sheep /store/$i -z $i -p 700$i; sleep 1; done
./collie/collie cluster format
for i in 2 3 4; do
pkill -f "sheep /store/$((i - 2))"
./sheep/sheep /store/$i -z $i -p 700$i
sleep 1
done
for i in 3 4; do pkill -f "sheep /store/$i"; sleep 1; done
for i in 0 1 2 3 4; do ./sheep/sheep /store/$i -z $i -p 700$i; sleep 1; done # now master 4 is recovered
for i in 0 1 2 3; do ./sheep/sheep /store/$i -z $i -p 700$i; sleep 1; done
for i in 0 1 2 3 4; do ./collie/collie cluster info -p 700$i; done

Cluster status: running

Creation time        Epoch Nodes
2011-09-25 11:47:49     12 [192.168.0.4:7000, 192.168.0.4:7001, 192.168.0.4:7002, 192.168.0.4:7003, 192.168.0.4:7004]
2011-09-25 11:47:49     11 [192.168.0.4:7000, 192.168.0.4:7001, 192.168.0.4:7002, 192.168.0.4:7004]
2011-09-25 11:47:49     10 [192.168.0.4:7000, 192.168.0.4:7001, 192.168.0.4:7004]
2011-09-25 11:47:49      9 [192.168.0.4:7000, 192.168.0.4:7004]
2011-09-25 11:47:49      8 [192.168.0.4:7004]
2011-09-25 11:47:49      7 [192.168.0.4:7003, 192.168.0.4:7004]
2011-09-25 11:47:49      6 [192.168.0.4:7003]
2011-09-25 11:47:49      5 [192.168.0.4:7002, 192.168.0.4:7003]
...

All is done with good and consistent epoch histories as expected.

Signed-off-by: Liu Yuan <tailai.ly@taobao.com>
Signed-off-by: MORITA Kazutaka <morita.kazutaka@lab.ntt.co.jp>