add vdi creation time to the vdi attribute header
Currently, Sheepdog doesn't clean up vdi attributes when we delete the
vdi.  So if we create a new vdi whose name was used before, we could
read the old attribute.

This patch adds a vdi creation time to the attribute header and checks
whether the attribute belongs to the current vdi.  Note that this is a
transitional approach. In future, we should remove vdi attributes when
we remove the vdi.

Signed-off-by: MORITA Kazutaka <morita.kazutaka@lab.ntt.co.jp>