sheep: move node membership management into cluster driver
Currently, Sheepdog has two node lists; sd_node_list and
cpg_node_list.  The former is used for consistent hashing and seen
from users.  The latter is managed in the cluster driver and notified
in join_handler/leave_handler.  But this design is too complex.  We
should move all the cluster management stuff into the cluster driver.
Main changes of this patch are as follows:

 - make join process one phase

   Node joining was really complex; cpg_confchg() notifies the newly
   joining node, the node multicasts a SD_MSG_JOIN message, and the
   master node receives it and multicasts the response.  Moreover, we
   couldn't allow any I/O events during two multicasting.  This patch
   moves all of them into the cluster driver.

 - add check_join_cb() to the join_handler() arguments

   This callback is called on one of the Sheepdog nodes (e.g. in the
   case of the corosync driver, the master server will call this).
   check_join_cb() checks whether the joining node may join the
   cluster, and returns the result.

 - use sheepdog_node_list_entry in the arguments of
   join_handler()/leave_handler()

   We can use the notified node list for consistent hashing now.

Signed-off-by: MORITA Kazutaka <morita.kazutaka@lab.ntt.co.jp>