sheep: remove node_list
join_handler() and leave_handler() can notify the node list which can
be used for consistent hashing now, so we don't need to manage node
lists in sheep/group.c any more.

With this patch, 'collie node list' doesn't show the master node.  But
I think it is okay because there may be no master node if we use the
other cluster driver than corosync.

Signed-off-by: MORITA Kazutaka <morita.kazutaka@lab.ntt.co.jp>