store: use fallocate when allocating new objects
Writing zeroes into the last sector of an object is not going to
preallocate it, but just allocates the last sector.  This leads
to fairly nasty fragmentation.  Use fallocate on the whole object
instead.  On my test setup with XFS this speeds up writes to an
unallocate volume from ~73MB/s to ~80MB/s.

If the filesystem does not support fallocate we fall back to the
old code.

Signed-off-by: Christoph Hellwig <hch@lst.de>
Signed-off-by: MORITA Kazutaka <morita.kazutaka@lab.ntt.co.jp>