sheep: abstract out store IO interface
We need to abstract out store IO interface to adopt it to other IO store
such as the coming 'Farm' store.

The open/read/write/close is cumbersome for a common kv-store to work with,
but this interface request smallest changes to current sheep store code. It
sucks but works as a kludge.

Don't get me wrong that I am writing an universal interface that will work well
with different kinds of data stores, say, sql-store, non-sql store, unstructured store,
the store that is not with local backing stroage, etc.

Simply I am *not* and I am always lost to foresee the future.

This interface is stupid but simply enough that costs me smallest changes to existing code
to let Sheepdog work with current store implementation and the coming 'Farm' store.

I think those kind people who try to squeeze other useful stores into Sheepdog are at a better
position to cook a more generic interface in the future.

- Why include length, offset that many kv stores don't need at all?

Okay, we'er trying to implement huge data size, so we need these to do partial object read/write.

- Why 'int fd' instead of a void *opaque for store object handle?

I suppose file is everything in UNIX philosophy and so fd can name everything and I hate type
conversion and frown when I can't cscope what it means for one second.

And last, I am happy to see anybody prove me wrong and replace it with a more capable interface.

Signed-off-by: Liu Yuan <tailai.ly@taobao.com>
Signed-off-by: MORITA Kazutaka <morita.kazutaka@lab.ntt.co.jp>