collie: use gateway for read I/O requests
Currently, collie sends read I/O requests without using gateway node
but calculating the target node as the sheep daemon does internally.
This is a hack to keep consistency even if the VM sends write requests
to the same objects at the same time.  But this enforces us to retry
the collie command manually when it fails due to node membership
changes.

This patch uses a gateway node for collie's read I/Os, and makes them
like QEMU's ones.  Data consistency of QEMU is not a problem if we use
SD_FLAG_CMD_WEAK_CONSISTENCY for collie's I/Os.

Signed-off-by: MORITA Kazutaka <morita.kazutaka@lab.ntt.co.jp>