farm: avoid unnecessary IO operation when recovering
When the cluster is recovering, we only need to write the objects
which no longer belong to the node to the snapshot, instead of writing
all the objects, by which we decreased the IO operation.

When we try to read an object, we first read it from the local
object directory, if not found, then read it from the snapshot.

Signed-off-by: levin li <xingke.lwp@taobao.com>
Signed-off-by: Liu Yuan <tailai.ly@taobao.com>