sheep: reduce snapshot COW read/write
Doing snapshot COW:

1, If new writing request occurs and need COW for sanpshot, now read old
object to buffer, then write the buffer to new object, then write the
request data to new object. We can merge the latter two writing request.

2, If new writing request covers whole object, no need to read old
object.

After the modification, pass bigger buffer to do_write_obj when doing
COW, but it will not add the burden. COW is never for inode object, so
it will not use the journal.

Signed-off-by: HaiTing Yao <wujue.yht@taobao.com>
Signed-off-by: Liu Yuan <tailai.ly@taobao.com>