collie: read the cached objects if any for collie operation
Collie operation such 'vdi list' or 'vdi object' need to read the lateset
VDI object, it might be cached in the object cache. We can't pass
SD_FLAG_CMD_CACHE to sheep because collie doesn't know if the cache mode is
enbled.

So we rely on the assumption that the cached object is freshest.

Signed-off-by: Liu Yuan <tailai.ly@taobao.com>