object cache: cache create operation
Currently create operation, we write through the cache. This operation would be
slow and return-err especially in node changes phase.

It is clean to let gateway only talks to object cache, and rely object cache layer
on pulling and pushing objects from/to sheep cluster.

With all the gateway requests operated on object cache to boost the IO performance,
we can do interested tricks with Farm, such as data de-duplication, which trade cpu
cycles for disk space while not hurting the IO performance.

Signed-off-by: Liu Yuan <tailai.ly@taobao.com>