logger: rotate log by size, default is 500MB
a simple patch to split log:  when log size exceed 500MB, a new
sheep.log file will be created, and old log will be renamed with
specified suffix(actually, it's the time when split happened).

the main purpose of this patch is to prevent log became too big,
which is hard to open and read.

Signed-off-by: Andy Chen <sanjie.cyg@taobao.com>
Signed-off-by: Yibin Shen <zituan@taobao.com>
Signed-off-by: MORITA Kazutaka <morita.kazutaka@lab.ntt.co.jp>
Signed-off-by: Liu Yuan <tailai.ly@taobao.com>