collie: make some optimization for parse_objs()
Currently, parse_objs traverse all the nodes to get the object,
and then call the callback function, it's OK for do_print_obj(),
but it's a waste for get_data_oid(), when the cluster gets larger
with more nodes, it makes collie slower the get the object id.

I add a return value for the callback function obj_parse_func_t,
if the function returns 1, the loop breaks when the first time
it succeeds, if 0, it will traverse the whole node list.

Signed-off-by: levin li <xingke.lwp@taobao.com>
Signed-off-by: Liu Yuan <tailai.ly@taobao.com>