deleting data objects of a vdi before deleting the inode
Currently, when deleting a vdi, sheep firstly clears the name
in the vdi inode to mark it as deleted, then try to delete the
data objects, there's a problem, if deleting one or more data
objects fails, we could never delete the object any more, because
the inode has been deleted.

Now I exchange the order of deleting inode and data objects, if
deleting some data object successes, we clear the the correlative
flag in inode->data_vdi_id[], orelse we set the dw->delete_error to
tell that an error occurs in the deletion work.

In delete_one_done(), if dw->delete_error is true, we set inode->vdi_size
to 0 to show this vdi has been deleted, but fail to delete some of
its objects, then we can try to delete the vdi again.

Signed-off-by: levin li <xingke.lwp@taobao.com>
Signed-off-by: Liu Yuan <tailai.ly@taobao.com>