always delete data objects when deleting an cloned vdi
When deleting a cloned vdi, sheep find the root vdi and then
traverse the vdi chain(such as base --> snapshot --> clone) to
check wheter there's an undeleted vdi in the chain, if some vdi
in the chain isn't deleted, sheep just mark the cloned vdi as
deleted by clear its vdi name.

But in fact a cloned vdi may created its own objects by copy-on-write,
these objects can be deleted when deleting the vdi, so we make
the cloned vdi to be deleted as the root vdi, then we can deleting
its data objects, in delete_one() we check whether the object belongs
to itself to determine whether to delete the object.

Signed-off-by: levin li <xingke.lwp@taobao.com>
Signed-off-by: Liu Yuan <tailai.ly@taobao.com>