sheep: cleanup nr_copies handling
Add a new get_nr_copies helper to consolidate calculating the number of
copies of an object that we need to deal with.  As a side effect this
allows making struct vnode_info private to group.c again.

There are a few places that take a different number of copies from the
inode or the on the wire header.  These look incorrect to me, but I'd
like to have some review of those.  These places should be either
documented or removed ASAP.

Signed-off-by: Christoph Hellwig <hch@lst.de>
Signed-off-by: Liu Yuan <tailai.ly@taobao.com>