logger: fix pid confusion
We need to store the logger (child) pid in a variable as getppid() fails
inside the death handler.  Instead of going back to using the pid variable
like it was done before "logger: improve death messages" use two different
variable for the sheep and logger, and comment on why we perform this
confusing dance, so that it doesn't get accidentally removed again.

Signed-off-by: Christoph Hellwig <hch@lst.de>
Signed-off-by: MORITA Kazutaka <morita.kazutaka@lab.ntt.co.jp>