sheep: add support for gateway mode
The gateway nodes only forwards requests to the sheepdog cluster, and
doesn't store objects in local at all.  Running a gateway daemon on
localhost and connecting to it would be an alternative to implementing
sheepdog fail-over support in qemu because the gateway node
automatically choose the correct target nodes.

Currently, this options is just an alias for '--vnodes 0'.  Note that
a small storage to store epoch information is still required for
gateway nodes.

Signed-off-by: MORITA Kazutaka <morita.kazutaka@gmail.com>