fix a bug of deleting base vdi fail
Take a view of the following snapshot tree:

base vdi --> snapshot vdi --> cloned vdi

when cloned vdi has its own data objects created by copy-on-write,
we firstly delete the cloned VDI, then delete the base VDI, at last
we delete snapshot VDI, the the snapshot VDI would delete fail in this
case because in the old code, I try to delete all the VDIs from the
VDI tree, but the base VDI and the cloned VDI have beed deleted,
and it causes an error which finally cause snapshot VDI delete fail.

This patch is another version of my previous one which is here:
http://lists.wpkg.org/pipermail/sheepdog/2012-May/003332.html

I simplely reset the delete_error flag, as Kazutaka mentions, it's
not good to do this, but we should never try to delete the VDIs
already deleted again, so I mark an VDI as deeply deleted by clearing
its name and size, which means the VDI and it data objects have already
been deleted, then in the next deletion work, we can just ignore these
VDIs. If the data objects haven't been deleted, then we just clear its
name to mark it as deleted, next time, we can try to delete its data
objects.

Signed-off-by: levin li <xingke.lwp@taobao.com>
Signed-off-by: Liu Yuan <tailai.ly@taobao.com>