sheep: need not read many copies of bitmap
Node maybe need read all of the bitmaps when the epoch is not zero. This
perhaps occurs when we restart the shutdown cluster or join one once
existed node.

Usually, node need not read many copies of bitmap. If the cluster status
is OK and the node status is waiting for format, node just need read one
bitamp.

Signed-off-by: HaiTing Yao <wujue.yht@taobao.com>
Signed-off-by: Liu Yuan <tailai.ly@taobao.com>