don't check for liburcu pkg-config files
For now we only use the uatomic.h header from liburcu and don't need
to link against the actual library, so we don't bother with the
PKG_CHECK_MODULES check.  If an actual dependency is introduced it
will need to be added back, including a non-pkgconfig check for older
distributions Like Debian -testing where the current check fails.

Signed-off-by: Christoph Hellwig <hch@lst.de>
Signed-off-by: Liu Yuan <tailai.ly@taobao.com>