sheep: add an option to disable object cache
It is highly recommended to read the document carefully before using
the object cache.

  https://github.com/collie/sheepdog/wiki/Backend-Stores-and-Object-Cache

This patch is useful for users whose environments don't meet the
requirements (e.g. sheep requires additional store on the gateway
nodes, we shouldn't access the same image from the different nodes
even if it is not the same time, etc).  If you are unsure how the
object cache works, it's safe to disable it.

In future, it's better to support multiple cache features because the
requirements for the sheepdog write cache seem to be quite different
among users.

Signed-off-by: MORITA Kazutaka <morita.kazutaka@lab.ntt.co.jp>