sheep: handle master crashing before sending join request
This patch fixes one of the cases described by Huxinwei and Liu Yuan where
sheepdog fails to elect a master. A longer description is in the patch.

A problem arises if a node joins the cluster and generates a confchg event,
then crashes or leaves without sending a join request and receiving a join
response. The second node to join never becomes master, and the entire cluster
hangs.

This patch allows a node to detect whether it should promote itself to master
after an arbitrary confchg event. Every node except the master creates a
blocked JOIN event for every node that joined after itself, therefore the
master is the node which has a JOIN event for every node in the members list.

Signed-off-by: Shevek <shevek@anarres.org>
Signed-off-by: Liu Yuan <tailai.ly@taobao.com>