sheep: pass vnode_info to VDI operations
There is no need to grab a local copy of the vnode information in the VDI ops,
as we already get it passed in through struct request.  For deletions we have
to grab an additional reference as deletions may still be in progress by the
time finish_request is called.

Signed-off-by: Christoph Hellwig <hch@lst.de>
Signed-off-by: Liu Yuan <tailai.ly@taobao.com>