sheep: queue cluster and local requests directly
Now that we can process events and requests in parallel, and events aren't
queued anymore we don't need to queue requests either.

Start with offloading local and cluster requests to the workqueues directly
in this patch and leave the slightly more complex handling of I/O requests
to the next patch.

Signed-off-by: Christoph Hellwig <hch@lst.de>
Signed-off-by: Liu Yuan <tailai.ly@taobao.com>