sheep: remove the request_queue
Offload I/O requests to the worker threads as soon as we get them in
queue_request, or when walking one of the lists of delayed requests.

Also move the code related to queueing I/O requests from group.c and
recovery.c into sdnet.c to have it in a single place.

Signed-off-by: Christoph Hellwig <hch@lst.de>
Signed-off-by: Liu Yuan <tailai.ly@taobao.com>