sheep: do not use strbuf for screen_obj_list
"sheep: add a helper function to copy out data from strbuf" added a new
strbuf_copyout helper that NULL terminates strings copied out of a strbuf.

For screen_objlist this leads to a off by one in the object list length and
recovery failures in many test cases.

Instead of simplify reverting that changes this patch changes screen_obj_list
to use a more efficienly algorithm for calculating the list of oids to be
replicated to the local node:

 - for each candidate OID we do a binary search on rw->oids for the number
   of elements before this call to screen_obj_list to replace the old call
   in merge_objlist.
 - then directly append the new oid to rw->oids and increment rw->count
 - after we are done processing the object list from one node we do a qsort
   pass over the oid list to make sure it will be sorted for the next
   iteration of screen_obj_list.

Signed-off-by: Christoph Hellwig <hch@lst.de>
Signed-off-by: Liu Yuan <tailai.ly@taobao.com>