sheep: fix local_get_node_list before the node joined a cluster
The GET_NODE_LIST command has the force flag and thus can be executed
before the cluster has started, which includes the short time before
the sheep even has any node list, in which we can trigger the assert
for current_vnode_info in get_vnode_info.

Fix this by checking for a NULL current_vnode_info first, for which we
have to move the implementation to group.c.  To me this makes sense
anyway as the node list is a group.c internal detail to start with.

Signed-off-by: Christoph Hellwig <hch@lst.de>
Signed-off-by: Liu Yuan <tailai.ly@taobao.com>