sheep: use requeue_request to retry requests in io_op_done
If we retry a request from io_op_done we currenly opencode the sequence in
queue_request, except that the call to check_request is missing.  Given
that the epoch may change the call to check_request is nesscary and we
should call queue_request in io_op_done as well.

Signed-off-by: Christoph Hellwig <hch@lst.de>
Signed-off-by: Liu Yuan <tailai.ly@taobao.com>